Split equivalence classes into smaller modules
